JavaScript/수업/day10/REST/requirements.txt
astroid==2.2.5
colorama==0.4.1
Django==2.2.1
django-bootstrap4==0.0.8
django-cors-headers==2.5.3
djangorestframework==3.9.3
isort==4.3.18
lazy-object-proxy==1.4.0
mccabe==0.6.1
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
wrapt==1.11.1
